pyproject.toml
[project]
name = "mcp-airflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv==1.1.1",
    "httpx==0.28.1",
    "mcp[cli]==1.12.2",
    "psycopg2-binary==2.9.10",
    "requests==2.32.4",
    "uvicorn==0.35.0",
]
